First check in for my AP project
The purpose of the program is to allow two human players to play a game of tic tac toe. When any of the nine cells are clicked, they are either marked with “X” or “O”. The program will check if the cell is already occupied. If it has been occupied, it will not be marked again. The turn also alternates as each player makes their move. In the end, when the game has been won or tied, it shows which player won the game (or shows that it was a tie), and the header and bottom text rapidly flash colors for 100,000 times if there is a winner and 50,000 times if it ended in a tie.